Move record-layer code to dedicated modules
The parts implementing the normal TLS record layer are moved to new
modules.  This adds a clear separation with the rest of the code in
upper layers.
