+++ core/Network/TLS/Sending.hs
@@ -10,12 +10,9 @@
 --
 module Network.TLS.Sending (
     encodePacket
-  , encodeRecord
-  , encodeRecordM
   , updateHandshake
   ) where
 
-import Network.TLS.Cap
 import Network.TLS.Cipher
 import Network.TLS.Context.Internal
 import Network.TLS.Handshake.State
@@ -57,31 +54,6 @@ packetToFragments ctx len (Handshake hss)  =
 packetToFragments _   _   (Alert a)        = return [encodeAlerts a]
 packetToFragments _   _   ChangeCipherSpec = return [encodeChangeCipherSpec]
 packetToFragments _   _   (AppData x)      = return [x]
-
--- before TLS 1.1, the block cipher IV is made of the residual of the previous block,
--- so we use cstIV as is, however in other case we generate an explicit IV
-prepareRecord :: Context -> RecordM a -> IO (Either TLSError a)
-prepareRecord ctx f = do
-    ver     <- usingState_ ctx (getVersionWithDefault $ maximum $ supportedVersions $ ctxSupported ctx)
-    txState <- readMVar $ ctxTxState ctx
-    let sz = case stCipher txState of
-                  Nothing     -> 0
-                  Just cipher -> if hasRecordIV $ bulkF $ cipherBulk cipher
-                                    then bulkIVSize $ cipherBulk cipher
-                                    else 0 -- to not generate IV
-    if hasExplicitBlockIV ver && sz > 0
-        then do newIV <- getStateRNG ctx sz
-                runTxState ctx (modify (setRecordIV newIV) >> f)
-        else runTxState ctx f
-
-encodeRecord :: Context -> Record Plaintext -> IO (Either TLSError ByteString)
-encodeRecord ctx = prepareRecord ctx . encodeRecordM
-
-encodeRecordM :: Record Plaintext -> RecordM ByteString
-encodeRecordM record = do
-    erecord <- engageRecord record
-    let (hdr, content) = recordToRaw erecord
-    return $ B.concat [ encodeHeader hdr, content ]
 
 switchTxEncryption :: Context -> IO ()
 switchTxEncryption ctx = do
